Changed type of Id in BaseEntity from ulong to int

+++ Repository/Interfaces/Generic/IGenericRepositoryAsync.cs
@@ -10,10 +10,10 @@ namespace Snippet.Data.Interfaces.Generic
         where TEntity : BaseEntity
     {
         Task<IReadOnlyCollection<TEntity>> GetAllAsync(CancellationToken ct);
-        Task<TEntity> GetByIdAsync(ulong id, CancellationToken ct);
+        Task<TEntity> GetByIdAsync(int id, CancellationToken ct);
         Task<IReadOnlyCollection<TEntity>> FindAsync(Func<TEntity, bool> predicate, CancellationToken ct);
         Task<TEntity> CreateAsync(TEntity entity, CancellationToken ct);
-        Task<bool> DeleteAsync(ulong id, CancellationToken ct);
+        Task<bool> DeleteAsync(int id, CancellationToken ct);
         Task<TEntity> UpdateAsync(TEntity entity, CancellationToken ct);
     }
 }
